+++ manager/working.go
@@ -1,6 +1,7 @@
 package manager
 
 import (
+	"context"
 	"encoding/json"
 	"time"
 
@@ -23,6 +24,14 @@ type Reservation struct {
 	Wid     string      `json:"wid"`
 	tsince  time.Time
 	texpiry time.Time
+}
+
+func (res *Reservation) ReservedAt() time.Time {
+	return res.tsince
+}
+
+func (res *Reservation) ExpiresAt() time.Time {
+	return res.texpiry
 }
 
 func (m *manager) WorkingCount() int {
@@ -84,13 +93,13 @@ func (m *manager) reserve(wid string, job *client.Job) error {
 	}
 
 	if timeout < 60 {
-		timeout = DefaultTimeout
 		util.Warnf("Timeout too short %d, 60 seconds minimum", timeout)
+		timeout = DefaultTimeout
 	}
 
 	if timeout > 86400 {
-		timeout = DefaultTimeout
 		util.Warnf("Timeout too long %d, one day maximum", timeout)
+		timeout = DefaultTimeout
 	}
 
 	exp := now.Add(time.Duration(timeout) * time.Second)
@@ -120,34 +129,27 @@ func (m *manager) reserve(wid string, job *client.Job) error {
 	return nil
 }
 
-func (m *manager) ack(jid string) (*client.Job, error) {
+func (m *manager) Acknowledge(jid string) (*client.Job, error) {
 	res := m.clearReservation(jid)
 	if res == nil {
 		util.Infof("No such job to acknowledge %s", jid)
 		return nil, nil
 	}
 
-	ok, err := m.store.Working().RemoveElement(res.Expiry, jid)
-	if !ok {
-		// doesn't matter, might not have acknowledged in time
-	}
-	return res.Job, err
-}
-
-func (m *manager) Acknowledge(jid string) (*client.Job, error) {
-	job, err := m.ack(jid)
+	// doesn't matter, might not have acknowledged in time
+	_, err := m.store.Working().RemoveElement(res.Expiry, jid)
 	if err != nil {
 		return nil, err
 	}
 
-	if job != nil {
+	if res.Job != nil {
 		m.store.Success()
-		err = callMiddleware(m.ackChain, job, func() error {
+		err = callMiddleware(m.ackChain, Ctx{context.Background(), res.Job, m, res}, func() error {
 			return nil
 		})
 	}
 
-	return job, err
+	return res.Job, err
 }
 
 func (m *manager) ReapExpiredJobs(timestamp string) (int, error) {